chore(vm) Extract `InstanceRef` into its own module.
